+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Feature
+
+- Add Finney Modal [#782](https://github.com/MyEtherWallet/MyEtherWallet/pull/782)
+
 ### Release v5.0.6
 
 ### Feature
@@ -9,7 +13,6 @@
 - Fix block 0 on mobile [#767](https://github.com/MyEtherWallet/MyEtherWallet/pull/767)
 - Update infura endpoint [#763](https://github.com/MyEtherWallet/MyEtherWallet/pull/763)
 - Chrome 72 compatibility for Secalot [#766](https://github.com/MyEtherWallet/MyEtherWallet/pull/766)
-
 
 ### Hotfix v5.0.5-hotfix.1